Move `typedEntries` and `typedKeys` to `advanced-types`
