+++ apps/hash-api/package.json
@@ -22,6 +22,7 @@
     "@blockprotocol/core": "0.0.12",
     "@blockprotocol/type-system": "0.0.3-canary-20230111084611",
     "@graphql-tools/schema": "8.5.0",
+    "@local/advanced-types": "0.0.0-private",
     "@local/hash-backend-utils": "0.0.0-private",
     "@local/hash-graph-client": "0.0.0-private",
     "@local/hash-graphql-shared": "0.0.0-private",
