+++ apps/hash-frontend/package.json
@@ -31,6 +31,7 @@
     "@glideapps/glide-data-grid": "5.2.1",
     "@hashintel/design-system": "0.0.1",
     "@lit-labs/react": "1.0.6",
+    "@local/advanced-types": "0.0.0-private",
     "@local/hash-graphql-shared": "0.0.0-private",
     "@local/hash-isomorphic-utils": "0.0.0-private",
     "@local/hash-subgraph": "0.0.0-private",
