+++ apps/hash-frontend/src/components/grid/utils/interactable-manager.ts
@@ -1,6 +1,6 @@
 import { CustomCell } from "@glideapps/glide-data-grid";
 import type { DrawArgs } from "@glideapps/glide-data-grid/dist/ts/data-grid/cells/cell-types";
-import { typedKeys } from "@local/hash-isomorphic-utils/util";
+import { typedKeys } from "@local/advanced-types/typed-entries";
 
 import {
   CellPath,
